chore: update breaking change doc, add tests
